Adding ember-cli-tailwind, ember-concurrency, ember-cli-mirage & ember-font-awesome
This package will be used for styling the application. There is a lot
included by default but builds can be trimmed down in the
confirguration.

This package is used to handle async behaviour in a way that won't
require excessive checking durning setup and teardown hooks. It will
also allow us to handle optionally showing UI elements (such as loading
spinners) while performing those async operations.

This will be used to mock an api. In this case it will act as our actual
api in production.

This package provides a large number of helpful icons, in production
only the icons that we are using will be included.