Adding the index route
I could use the application route here however there are two reasons
that I shouldn't.

Firstly the spec requires that the component (customer-search) must
accept a function passed to it. This means I need to have a controller
and that would have to be the application controller, which leads into
reason two.

> I could use ember-route-actions here but that seems very excessive for
a single action

Secondly the application route (and hence it's template & controller)
are parent elements that wrap all other components/routes. As such their
logic is run every time the application is loaded. If we were to expand
the application we likely would not want the customer search logic to
run or at the very least pollute the namespace.

run or at the very least polute the namespace.