Adding keyboard navigation and customer selection/clearing
Part one of this will add the customers Last and First names to the
button inplace of the input field. Part two will be showing a tooltip
and allowing the user to select the customer for editing.

Allows the user to move up and down the list of customers using their
arrow keys and allows enter to be used to select a customer