Adding ember modal dialog
This is a handy package for creating/showing modals within your
application. We could also add ember-tether and create our own modals
from scratch and depending on how configurable the modal needs to be
that may be the better option. However for a project of this size
and purpose there's no need to re-implement a well thought out package.