Initial setup for showing & hiding the customer form modal
As per the requirements (and I believe good practice in this case) I
will be passing down two actions to the cusotmer-search component. The
first action will be used to create a new customer and open the customer
details form. The later will take a customer and poulate their details
into the customer details form for editing.