Index Controller: Using a constant for the modal, adding saveCustomer task and rolling back any unsaved changes when the modal closes
Using a constant allows us to change the underlying string later if
needed.

Persist a new customer or the changes we have made to an existing
customer to the mirage database.

If the user closes the modal we want them to rollback any changes that
they may have made. This ensures that the state of the app does not fall
out of sync with the database or surprise the user.