Providing some tooltips for our customer search component via ember-tooltips
This is a super handy library that is well used in the ember eco-system
for handling tooltips. Since tailwind does not provide a tooltips
component we can avoid building our own via this package.

We could also use this packages custom styling to provide more detailed
guidance to the user with the customer-details form if a validation is
incorrect.