Adding the ember-cp-validations package for validating the customer form
This is a fantastic package that allows you to define a set
of validations for any model or piece of data. Using there validator
syntax (which I have stored under lib) you can easily compose
validations, setup debounce timings, create custom messages and really
get your form validation working the way you want it.

Using this package it makes it really easy to extend the validations on
our controller, and as we will see in the next commit add validation
messages and styling to our UiInput component.